advanced nested loops

+++ functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int n = 50;
-	long long int fib_sequence[n];
+	int fib_sequence[n];
 
 	fib_sequence[0] = 1;
 	fib_sequence[1] = 2;
@@ -21,7 +21,7 @@ int main(void)
 
 	for (int i = 0; i < n; i++)
 	{
-		printf("%lld", fib_sequence[i]);
+		printf("%d", fib_sequence[i]);
 		if (i < n - 1)
 		{
 			printf(", ");
